feat: moved common elements to react-lib
