+++ src/pages/pools/Pool/CreatePool.tsx
@@ -7,7 +7,6 @@ import { network as libNet } from '@reef-chain/util-lib';
 import TokenContext from '../../../context/TokenContext';
 import TokenPricesContext from '../../../context/TokenPricesContext';
 import { notify } from '../../../utils/utils';
-import '../../../common/overlay-swap.css';
 import './create-pool.css';
 import { localizedStrings } from '../../../l10n/l10n';
 import ReefSigners from '../../../context/ReefSigners';
